Refactor error handling
